Add tiny movement code for physics testing
Collision and physics systems require some user movement to check if
they're playing nice and working as intended. So, basic movement on
sprite0 was created for demoing said systems.

Although movement code is game specific, and thus shouldn't live in the
engine, a prototype is accessible via the DEBUG_MODE flag, or `make
debug`. It will need to be decoupled from sprite later on, when
GameObjects exist.

changelog:
- add basic move left, right, up, down functions on sprite0
- add keyboard event listener for arrow keys to eon.c
- add movement.o target to Makefile
- add PHONY debug target to Makefile to invoke movement.o linking